requirements.txt
streamlit==1.28.1
gitpython==3.1.40
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
altair==5.1.2
seaborn==0.12.2
matplotlib==3.8.2
python-dateutil==2.8.2
